refactor(type-checker): extract Helpers partial class [slice 1/6]
Moves 29 stateless utility methods from PreceptTypeChecker.cs into
PreceptTypeChecker.Helpers.cs. Adds 'partial' keyword to the main
class declaration.

No behavior changes, no signature changes, no visibility changes.
1752 Precept.Tests pass.

Closes #118 (partial)
